fix 105No1,4,5

+++ Works/105No1Step4.c
@@ -2,7 +2,7 @@
 #include<conio.h>
 
 int table;
-int price;
+float price;
 float sum;
 
 int main ()
@@ -10,7 +10,7 @@ int main ()
     printf("Enter the number of the tables :");
     scanf("%d", &table);
     printf("\nEnter the cost per table : ");
-    scanf("%d",&price);
+    scanf("%f",&price);
     sum = table*price;
     printf("\nTotal cost : %.2f",sum);
     getch();
